refactor: Move theme code from utils package into app.
