+++ apps/www/app/root.tsx
@@ -12,15 +12,13 @@ import {
 import type { LoaderFunctionArgs } from '@remix-run/node';
 
 import { RouterProvider } from '@tipprunde/ui';
-import {
-  MediaQueryFallback,
-  ThemeProvider,
-  getHints,
-  useTheme,
-} from '@tipprunde/utils/theme';
 
 import { getUser } from '#app/.server/auth';
+import { getHints } from '#app/.server/client-hints';
 import { getSession } from '#app/.server/theme';
+
+import { MediaQueryFallback } from '#app/utils/media-query-fallback';
+import { ThemeProvider, useTheme } from '#app/utils/theme';
 
 import './styles/tailwind.css';
