+++ apps/www/app/utils/media-query-fallback.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from './theme-provider';
+import { useTheme } from './theme';
 
 const clientThemeCode = `
 (() => {
