+++ packages/utils/package.json
@@ -2,8 +2,7 @@
   "name": "@tipprunde/utils",
   "version": "0.3.0",
   "exports": {
-    ".": "./src/index.ts",
-    "./theme": "./src/theme/index.ts"
+    ".": "./src/index.ts"
   },
   "devDependencies": {
     "@remix-run/node": "^2.6.0",
