--- packages/utils/src/theme/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-export * from './client-hints';
-export * from './media-query-fallback';
-export * from './theme-provider';
