--- packages/utils/src/theme/types.ts
@@ -1,22 +1,0 @@
-import * as v from 'valibot';
-
-const colorSchemeNames = ['light', 'dark'] as const;
-export const ColorSchemeSchema = v.picklist(colorSchemeNames);
-
-export const ThemeSchema = v.object({
-  colorScheme: ColorSchemeSchema,
-  // themeColor: ...
-});
-export const ClientHintsSchema = v.object({
-  colorScheme: v.nullable(ColorSchemeSchema),
-});
-
-export type ColorSchemeSource = 'client' | 'session';
-export type ColorScheme = v.Input<typeof ColorSchemeSchema>;
-
-export type Theme = v.Input<typeof ThemeSchema>;
-export type ThemeSessionData = {
-  theme: Theme;
-};
-
-export type ClientHints = v.Input<typeof ClientHintsSchema>;
